streamlit/requirements.txt
streamlit>=1.28.0
groq>=0.4.0
anthropic>=0.18.0
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
